# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "User Game API",
        "version": "1.1.0",
        "description": "Simple game api that store user account, biodata, and game histories"
    },
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Example: Bearer [JWT TOKEN]"
        }
    },
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "basePath": "/api/v1",
...
    "paths": {
...
        "/user_games/biodatas": {
            "get": {
                "summary": "Get all user game biodatas",
                "operationId": "GetUserGameBiodatas",
                "responses": {
                    "200": {
                        "description": "Returns a list of user game biodatas",
                        "schema": {
                            "$ref": "#/responses/getUserGameBiodatasResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "tags": [
                    "User Game Biodata"
                ]
            },
            "post": {
                "summary": "Create a new user game biodata",
                "operationId": "CreateUserGameBiodatas",
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns user game biodata that has been created",
                        "schema": {
                            "$ref": "#/responses/postUserGameBiodatasResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/responses/badRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "tags": [
                    "User Game Biodata"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "email"
                    },
                    {
                        "name": "firstname",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "lastname",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "profilePicture",
                        "in": "formData",
                        "required": true,
                        "type": "file"
                    },
                    {
                        "name": "country",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "age",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    {
                        "name": "userGameId",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    }
                ]
            }
        },
        "/user_game/biodata/{id}": {
            "get": {
                "summary": "Get user game biodata by id",
                "operationId": "GetUserGameBiodata",
                "responses": {
                    "200": {
                        "description": "Returns a user game biodata",
                        "schema": {
                            "$ref": "#/responses/getUserGameBiodataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/responses/badRequestResponse"
                        }
                    },
                    "404": {
                        "description": "User game biodata not found",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User game biodata id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "User Game Biodata"
                ]
            },
            "patch": {
                "summary": "Update user game biodata by id",
                "operationId": "UpdateUserGameBiodata",
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns before after user game biodata field that has been updated",
                        "schema": {
                            "$ref": "#/responses/patchUserGameBiodataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/responses/badRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User game biodata not found",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "tags": [
                    "User Game Biodata"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User game biodata id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "required": false,
                        "type": "string",
                        "format": "email"
                    },
                    {
                        "name": "firstname",
                        "in": "formData",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "lastname",
                        "in": "formData",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "profilePicture",
                        "in": "formData",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "name": "country",
                        "in": "formData",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "age",
                        "in": "formData",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    {
                        "name": "userGameId",
                        "in": "formData",
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    }
                ]
            },
            "delete": {
                "summary": "Delete user game biodata by id",
                "operationId": "DeleteUserGameBiodata",
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns user game biodata that has been deleted",
                        "schema": {
                            "$ref": "#/responses/deleteUserGameBiodataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/responses/badRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User game biodata not found",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "tags": [
                    "User Game Biodata"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User game biodata id",
                        "required": true,
                        "type": "integer"
                    }
                ]
            }
        },
...
    },
...
    "responses": {
...
        "getUserGameBiodatasResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 200
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "OK"
                },
                "count": {
                    "type": "integer",
                    "format": "int32"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/userGameBiodataInclude"
                    }
                }
            }
        },
        "postUserGameBiodatasResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 201
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "UserGameBiodata created successfully"
                },
                "data": {
                    "$ref": "#/definitions/userGameBiodata"
                }
            }
        },
        "getUserGameBiodataResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 200
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "OK"
                },
                "data": {
                    "$ref": "#/definitions/userGameBiodataInclude"
                }
            }
        },
        "patchUserGameBiodataResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 200
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "UserGameBiodata updated successfully"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "before": {
                            "$ref": "#/requestBodies/patchUserGameBiodataRequest"
                        },
                        "after": {
                            "$ref": "#/requestBodies/patchUserGameBiodataRequest"
                        }
                    }
                }
            }
        },
        "deleteUserGameBiodataResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 200
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "UserGameBiodata deleted successfully"
                },
                "data": {
                    "$ref": "#/definitions/userGameBiodata"
                }
            }
        },
...
        "generalErrorResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string",
                    "format": "string"
                }
            }
        },
        "badRequestResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 400
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "value": {
                                "type": "string",
                                "format": "string"
                            },
                            "msg": {
                                "type": "string",
                                "format": "string"
                            },
                            "param": {
                                "type": "string",
                                "format": "string"
                            },
                            "location": {
                                "type": "string",
                                "format": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "userGame": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "username": {
                    "type": "string",
                    "format": "string"
                },
                "password": {
                    "type": "string",
                    "format": "string"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
        "userGameBiodata": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "email": {
                    "type": "string",
                    "format": "string"
                },
                "firstname": {
                    "type": "string",
                    "format": "string"
                },
                "lastname": {
                    "type": "string",
                    "format": "string"
                },
                "profilePicture": {
                    "type": "string",
                    "format": "string"
                },
                "country": {
                    "type": "string",
                    "format": "string"
                },
                "age": {
                    "type": "integer",
                    "format": "int32"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "userGameId": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
        "userGameBiodataInclude": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "email": {
                    "type": "string",
                    "format": "string"
                },
                "firstname": {
                    "type": "string",
                    "format": "string"
                },
                "lastname": {
                    "type": "string",
                    "format": "string"
                },
                "profilePicture": {
                    "type": "string",
                    "format": "string"
                },
                "country": {
                    "type": "string",
                    "format": "string"
                },
                "age": {
                    "type": "integer",
                    "format": "int32"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "userGameId": {
                    "type": "integer",
                    "format": "int32"
                },
                "UserGame": {
                    "$ref": "#/definitions/userGame"
                }
            }
        },
...
    }
}